move resources format to new file
